Refactor Phase 1: package scaffold + CLI shims
- Add pyproject.toml with project metadata, deps, and console_scripts
  (connect-migrate-discover, connect-migrate-migrate).
- Create src/connect_migrate/ package; move all src/*.py into it (flat
  layout for now; Phase 2 reorganizes into subdirs).
- Move entry-point scripts into src/connect_migrate/cli/{discover_cli,
  migrate_cli}.py.
- Dedupe setup_logging into src/connect_migrate/utils/logging_setup.py
  (was identical in both entry points).
- Replace src/discovery_script.py and src/migrate_connector_script.py
  with 3-line shims so existing customer commands still work:
    python src/discovery_script.py ...
    python src/migrate_connector_script.py ...
- Rewrite intra-package imports to use `from connect_migrate.X import Y`.

No logic changes. Customer-facing CLI flags, output paths, and JSON
schema unchanged.
