+++ src/connect_migrate/__init__.py
@@ -1,0 +1,9 @@
+"""Confluent Connect Migration Utility.
+
+Maps self-managed Kafka Connect connector configurations to Confluent Cloud
+fully-managed connector configurations, and creates the migrated connectors.
+
+Customer entry points live in :mod:`connect_migrate.cli`.
+"""
+
+__version__ = "0.1.0"
