+++ src/connect_migrate/utils/logging_setup.py
@@ -1,0 +1,26 @@
+import logging
+import sys
+from pathlib import Path
+
+
+def setup_logging(output_dir: Path) -> None:
+    """Configure root logger to write to ``<output_dir>/migration.log`` and stdout."""
+    log_file = output_dir / "migration.log"
+
+    root_logger = logging.getLogger()
+    for handler in root_logger.handlers[:]:
+        root_logger.removeHandler(handler)
+
+    formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+
+    file_handler = logging.FileHandler(log_file)
+    file_handler.setFormatter(formatter)
+    file_handler.setLevel(logging.INFO)
+
+    console_handler = logging.StreamHandler(sys.stdout)
+    console_handler.setFormatter(formatter)
+    console_handler.setLevel(logging.INFO)
+
+    root_logger.setLevel(logging.INFO)
+    root_logger.addHandler(file_handler)
+    root_logger.addHandler(console_handler)
